HR Agent App/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
streamlit>=1.31.0
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.4.0
langchain-classic>=0.1.0
langchain-chroma>=0.1.0
pydantic>=2.0.0,<3
python-dotenv>=1.0.0
mcp[cli]>=1.0.0
chromadb>=0.5.0
rank-bm25>=0.2.2
